factor out common initialization code to a parent struct
