+++ src/blsct/arith/mcl/init/atomic_mcl_init.h
@@ -7,11 +7,12 @@
 
 #define BLS_ETH 1
 #include <bls/bls384_256.h>
+#include <blsct/arith/mcl/init/mcl_init.h>
 #include <stdexcept>
 #include <boost/thread/lock_guard.hpp>
 #include <boost/thread/mutex.hpp>
 
-class AtomicMclInit
+class AtomicMclInit : MclInit 
 {
 public:
     AtomicMclInit()
@@ -20,10 +21,7 @@ public:
         static bool is_initialized = false;
         if (is_initialized) return;
 
-        if (blsInit(MCL_BLS12_381, MCLBN_COMPILED_TIME_VAR) != 0) {
-            throw std::runtime_error("blsInit failed");
-        }
-        mclBn_setETHserialization(1);
+        Initialize();
 
         is_initialized = true;
     }
