+++ src/test/util/setup_common.h
@@ -6,7 +6,7 @@
 #define BITCOIN_TEST_UTIL_SETUP_COMMON_H
 
 #include <blsct/arith/mcl/mcl_g1point.h>
-#include <blsct/arith/mcl/atomic_mcl_init.h>
+#include <blsct/arith/mcl/init/atomic_mcl_init.h>
 #include <chainparamsbase.h>
 #include <fs.h>
 #include <key.h>
